requirements.txt
# Core Dependencies (Required)
pymupdf==1.23.8
pdfplumber==0.10.3

# PII Redaction (Required)
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
spacy==3.7.2

# Optional (Enhanced Features)
# paddlepaddle==2.6.0
# paddleocr==2.7.0.3

# Supporting Libraries
Pillow==10.1.0
python-docx==1.1.0

# Note: After installing spacy, run:
# python -m spacy download en_core_web_sm
